pyproject.toml
[project]
name = "chatlog-2-hatena-web"
version = "0.1.0"
description = "AI-chatbot conversation summarizer and blog auto-poster on the web"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.6",
    "fastapi>=0.126.0",
    "google-genai>=1.56.0",
    "gspread>=6.2.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=2.14.0",
    "psycopg[binary]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.38.0",
    "yfinance>=0.2.66",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]

[project.urls]
Repository = "https://github.com/HiroItozzz/chatlog-2-hatena-web.git"

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
include = ["app*", "cha2hatena*"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80
